Conflicts:
	src/uCtrlDesktop/Resources.qrc
	src/uCtrlDesktop/qml/device/type/LightSensor.qml

+++ src/uCtrlCore/Protocol/uctrlapifacade.cpp
@@ -2,6 +2,7 @@
 
 UCtrlAPIFacade::UCtrlAPIFacade(QNetworkAccessManager* nam, UPlatformsModel* platforms, QObject *parent)
     : m_uCtrlApi(nam, platforms, parent)
+    , m_uCtrlLocalApi(platforms, parent)
 {
 }
 
@@ -56,58 +57,104 @@ void UCtrlAPIFacade::getPlatforms()
 
 void UCtrlAPIFacade::postPlatform(UPlatform* platform)
 {
+    if (platform->isLocalPlatform())
+        return;
+
     m_uCtrlApi.postPlatform(platform->id());
 }
 
 void UCtrlAPIFacade::getPlatform(UPlatform* platform)
 {
+    // Not working on uCtrl Hardware, platforms are broadcasted
+    if (platform->isLocalPlatform())
+        return;
+
     m_uCtrlApi.getPlatform(platform->id());
 }
 
 void UCtrlAPIFacade::putPlatform(UPlatform* platform)
 {
-    m_uCtrlApi.putPlatform(platform->id());
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.savePlatform(platform);
+    else
+        m_uCtrlApi.putPlatform(platform->id());
 }
 
 void UCtrlAPIFacade::deletePlatform(UPlatform* platform)
 {
-    m_uCtrlApi.deletePlatform(platform->id());
+    if (platform->isLocalPlatform()) // WHAT DO?
+        return;
+    else
+        m_uCtrlApi.deletePlatform(platform->id());
 }
 
 void UCtrlAPIFacade::getDevices(UPlatform* platform)
 {
-    m_uCtrlApi.getDevices(platform->id());
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.getDevices(platform);
+    else
+        m_uCtrlApi.getDevices(platform->id());
 }
 void UCtrlAPIFacade::postDevice(UDevice* device)
 {
-    QString platformId;
-    resolveIds(platformId, device);
-    m_uCtrlApi.postDevice(platformId, device->id());
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        putDevice(device);
+    else
+    {
+        QString platformId;
+        resolveIds(platformId, device);
+        m_uCtrlApi.postDevice(platformId, device->id());
+    }
 }
 void UCtrlAPIFacade::getDevice(UDevice* device)
 {
-    QString platformId;
-    resolveIds(platformId, device);
-    m_uCtrlApi.getDevice(platformId, device->id());
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+    {
+        // TODO
+        return;
+    }
+    else
+    {
+        QString platformId;
+        resolveIds(platformId, device);
+        m_uCtrlApi.getDevice(platformId, device->id());
+    }
 }
 void UCtrlAPIFacade::putDevice(UDevice* device)
 {
-    QString platformId;
-    resolveIds(platformId, device);
-    m_uCtrlApi.putDevice(platformId, device->id());
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.saveDevices(platform);
+    else
+    {
+        QString platformId;
+        resolveIds(platformId, device);
+        m_uCtrlApi.putDevice(platformId, device->id());
+    }
 }
 void UCtrlAPIFacade::deleteDevice(UDevice* device)
 {
-    QString platformId;
-    resolveIds(platformId, device);
-    m_uCtrlApi.deleteDevice(platformId, device->id());
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        putDevice(device);
+    else
+    {
+        QString platformId;
+        resolveIds(platformId, device);
+        m_uCtrlApi.deleteDevice(platformId, device->id());
+    }
 }
 
 void UCtrlAPIFacade::getDeviceAllStats(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
-    m_uCtrlApi.getDeviceValues(platformId, device->id(), params);
     m_uCtrlApi.getDeviceMax(platformId, device->id(), params);
     m_uCtrlApi.getDeviceMin(platformId, device->id(), params);
     m_uCtrlApi.getDeviceMean(platformId, device->id(), params);
@@ -116,6 +163,10 @@ void UCtrlAPIFacade::getDeviceAllStats(UDevice *device, QMap<QString, QVariant> 
 
 void UCtrlAPIFacade::getDeviceValues(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceValues(platformId, device->id(), params);
@@ -123,6 +174,10 @@ void UCtrlAPIFacade::getDeviceValues(UDevice *device, QMap<QString, QVariant> pa
 
 void UCtrlAPIFacade::getDeviceMax(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceMax(platformId, device->id(), params);
@@ -130,6 +185,10 @@ void UCtrlAPIFacade::getDeviceMax(UDevice *device, QMap<QString, QVariant> param
 
 void UCtrlAPIFacade::getDeviceMin(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceMin(platformId, device->id(), params);
@@ -137,6 +196,10 @@ void UCtrlAPIFacade::getDeviceMin(UDevice *device, QMap<QString, QVariant> param
 
 void UCtrlAPIFacade::getDeviceMean(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceMean(platformId, device->id(), params);
@@ -144,6 +207,10 @@ void UCtrlAPIFacade::getDeviceMean(UDevice *device, QMap<QString, QVariant> para
 
 void UCtrlAPIFacade::getDeviceCount(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceCount(platformId, device->id(), params);
@@ -151,6 +218,10 @@ void UCtrlAPIFacade::getDeviceCount(UDevice *device, QMap<QString, QVariant> par
 
 void UCtrlAPIFacade::getDeviceHistory(UDevice *device, QMap<QString, QVariant> params)
 {
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     resolveIds(platformId, device);
     m_uCtrlApi.getDeviceHistory(platformId, device->id(), params);
@@ -158,21 +229,37 @@ void UCtrlAPIFacade::getDeviceHistory(UDevice *device, QMap<QString, QVariant> p
 
 void UCtrlAPIFacade::getScenarios(UDevice* device)
 {
-    QString platformId;
-    resolveIds(platformId, device);
-    m_uCtrlApi.getScenarios(platformId, device->id());
+    UPlatform* platform = (UPlatform*)device->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.getScenarios(device);
+    else
+    {
+        QString platformId;
+        resolveIds(platformId, device);
+        m_uCtrlApi.getScenarios(platformId, device->id());
+    }
 }
 
 void UCtrlAPIFacade::postScenario(UScenario* scenario)
 {
-    QString platformId;
-    QString deviceId;
-    resolveIds(platformId, deviceId, scenario);
-    QString scenarioId = scenario->id();
-    m_uCtrlApi.postScenario(platformId, deviceId, scenario->id());
+    UPlatform* platform = (UPlatform*)scenario->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putScenario(scenario);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        resolveIds(platformId, deviceId, scenario);
+        m_uCtrlApi.postScenario(platformId, deviceId, scenario->id());
+    }
 }
+
 void UCtrlAPIFacade::getScenario(UScenario* scenario)
 {
+    UPlatform* platform = (UPlatform*)scenario->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     QString deviceId;
     resolveIds(platformId, deviceId, scenario);
@@ -180,38 +267,67 @@ void UCtrlAPIFacade::getScenario(UScenario* scenario)
 }
 void UCtrlAPIFacade::putScenario(UScenario* scenario)
 {
-    QString platformId;
-    QString deviceId;
-    resolveIds(platformId, deviceId, scenario);
-    m_uCtrlApi.putScenario(platformId, deviceId, scenario->id());
+    UDevice* device = (UDevice*)scenario->parent()->parent();
+    UPlatform* platform = (UPlatform*)device;
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.saveScenarios(device);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        resolveIds(platformId, deviceId, scenario);
+        m_uCtrlApi.putScenario(platformId, deviceId, scenario->id());
+    }
 }
 void UCtrlAPIFacade::deleteScenario(UScenario* scenario)
 {
-    QString platformId;
-    QString deviceId;
-    resolveIds(platformId, deviceId, scenario);
-    m_uCtrlApi.deleteScenario(platformId, deviceId, scenario->id());
+    UPlatform* platform = (UPlatform*)scenario->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putScenario(scenario);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        resolveIds(platformId, deviceId, scenario);
+        m_uCtrlApi.deleteScenario(platformId, deviceId, scenario->id());
+    }
 }
 
 void UCtrlAPIFacade::getTasks(UScenario* scenario)
 {
-    QString platformId;
-    QString deviceId;
-    resolveIds(platformId, deviceId, scenario);
-    m_uCtrlApi.getTasks(platformId, deviceId, scenario->id());
+    UPlatform* platform = (UPlatform*)scenario->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.getTasks(scenario);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        resolveIds(platformId, deviceId, scenario);
+        m_uCtrlApi.getTasks(platformId, deviceId, scenario->id());
+    }
 }
 
 void UCtrlAPIFacade::postTask(UTask* task)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    resolveIds(platformId, deviceId, scenarioId, task);
-    m_uCtrlApi.postTask(platformId, deviceId, scenarioId, task->id());
+    UPlatform* platform = (UPlatform*)task->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putTask(task);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        resolveIds(platformId, deviceId, scenarioId, task);
+        m_uCtrlApi.postTask(platformId, deviceId, scenarioId, task->id());
+    }
 }
 
 void UCtrlAPIFacade::getTask(UTask* task)
 {
+    UPlatform* platform = (UPlatform*)task->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     QString deviceId;
     QString scenarioId;
@@ -221,43 +337,72 @@ void UCtrlAPIFacade::getTask(UTask* task)
 
 void UCtrlAPIFacade::putTask(UTask* task)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    resolveIds(platformId, deviceId, scenarioId, task);
-    m_uCtrlApi.putTask(platformId, deviceId, scenarioId, task->id());
+    UPlatform* platform = (UPlatform*)task->parent()->parent()->parent()->parent()->parent()->parent();
+    UScenario* scenario = (UScenario*)task->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.saveTasks(scenario);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        resolveIds(platformId, deviceId, scenarioId, task);
+        m_uCtrlApi.putTask(platformId, deviceId, scenarioId, task->id());
+    }
 }
 
 void UCtrlAPIFacade::deleteTask(UTask* task)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    resolveIds(platformId, deviceId, scenarioId, task);
-    m_uCtrlApi.deleteTask(platformId, deviceId, scenarioId, task->id());
+    UPlatform* platform = (UPlatform*)task->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putTask(task);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        resolveIds(platformId, deviceId, scenarioId, task);
+        m_uCtrlApi.deleteTask(platformId, deviceId, scenarioId, task->id());
+    }
 }
 
 void UCtrlAPIFacade::getConditions(UTask* task)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    resolveIds(platformId, deviceId, scenarioId, task);
-    m_uCtrlApi.getConditions(platformId, deviceId, scenarioId, task->id());
+    UPlatform* platform = (UPlatform*)task->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.getConditions(task);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        resolveIds(platformId, deviceId, scenarioId, task);
+        m_uCtrlApi.getConditions(platformId, deviceId, scenarioId, task->id());
+    }
 }
 
 void UCtrlAPIFacade::postCondition(UCondition* condition)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    QString taskId;
-    resolveIds(platformId, deviceId, scenarioId, taskId, condition);
-    m_uCtrlApi.postCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    UPlatform* platform = (UPlatform*)condition->parent()->parent()->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putCondition(condition);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        QString taskId;
+        resolveIds(platformId, deviceId, scenarioId, taskId, condition);
+        m_uCtrlApi.postCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    }
 }
 
 void UCtrlAPIFacade::getCondition(UCondition* condition)
 {
+    UPlatform* platform = (UPlatform*)condition->parent()->parent()->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        return;
+
     QString platformId;
     QString deviceId;
     QString scenarioId;
@@ -268,22 +413,35 @@ void UCtrlAPIFacade::getCondition(UCondition* condition)
 
 void UCtrlAPIFacade::putCondition(UCondition* condition)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    QString taskId;
-    resolveIds(platformId, deviceId, scenarioId, taskId, condition);
-    m_uCtrlApi.putCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    UPlatform* platform = (UPlatform*)condition->parent()->parent()->parent()->parent()->parent()->parent()->parent()->parent();
+    UTask* task = (UTask*)condition->parent()->parent();
+    if (platform->isLocalPlatform())
+        m_uCtrlLocalApi.saveConditions(task);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        QString taskId;
+        resolveIds(platformId, deviceId, scenarioId, taskId, condition);
+        m_uCtrlApi.putCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    }
 }
 
 void UCtrlAPIFacade::deleteCondition(UCondition* condition)
 {
-    QString platformId;
-    QString deviceId;
-    QString scenarioId;
-    QString taskId;
-    resolveIds(platformId, deviceId, scenarioId, taskId, condition);
-    m_uCtrlApi.deleteCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    UPlatform* platform = (UPlatform*)condition->parent()->parent()->parent()->parent()->parent()->parent()->parent()->parent();
+    if (platform->isLocalPlatform())
+        putCondition(condition);
+    else
+    {
+        QString platformId;
+        QString deviceId;
+        QString scenarioId;
+        QString taskId;
+        resolveIds(platformId, deviceId, scenarioId, taskId, condition);
+        m_uCtrlApi.deleteCondition(platformId, deviceId, scenarioId, taskId, condition->id());
+    }
 }
 
 void UCtrlAPIFacade::getRecommendations()
